[JT-B-J-026] Changes have been made to the logic of the user ID deletion method. And also added unit testing for the method of deleting a user by ID. Three possible cases are considered: 1) Successful deletion of the user by ID. 2) The user with this ID has not been found. 3) Database connection error.

+++ auth/src/main/java/com/jointAuth/service/UserService.java
@@ -97,6 +97,11 @@ public class UserService {
     }
 
     public void deleteUser(Long id) {
-        userRepository.deleteById(id);
+        Optional<User> optionalUser = userRepository.findById(id);
+        if (optionalUser.isPresent()) {
+            userRepository.deleteById(id);
+        } else {
+            throw new IllegalArgumentException("User not found");
+        }
     }
 }
